backend/requirements.txt
# AegisTwin Backend — Python Dependencies
# For development: pip install -e ".[dev]" from the project root
# For manual install without editable mode: pip install -r backend/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
celery[redis]>=5.4.0
redis>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
structlog>=24.1.0
rich>=13.7.0
openai>=1.0.0
networkx>=3.3
